backend/requirements.txt
pandas>=2.2.3
requests>=2.32.5
beautifulsoup4>=4.14.2
great_expectations>=1.11.3
pytest>=9.0.2
flask>=3.0.3
flask-cors>=5.0.0
waitress>=3.0.2
python-dotenv>=1.0.1
python-docx>=1.1.2
openpyxl>=3.1.2
urllib3<2.0.0
Pillow>=11.0.0
pyarrow>=21.0.0
fastparquet
fastapi
uvicorn
python-multipart
ydata-profiling
scipy
reportlab>=4.3.1
supabase>=2.13.0
asyncpg>=0.30.0
httpx>=0.27.0
